Security roles

+++ src/AppBundle/Controller/ForumController.php
@@ -15,6 +15,7 @@ use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use AppBundle\Entity\Discussion;
 use AppBundle\Entity\Themes;
 use \DateTime;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 
 
 class ForumController extends Controller
@@ -130,15 +131,10 @@ class ForumController extends Controller
           ->getForm();
 
       $formAdd->handleRequest($request);
-
-      if ($formAdd->isSubmitted() && $formAdd->isValid()) {
+      $securityContext = $this->get('security.authorization_checker');
+      if ($formAdd->isSubmitted() && $formAdd->isValid() && $securityContext->isGranted('IS_AUTHENTICATED_FULLY')) {
         try {
-        // USE SESSION TO GET SESSION USER OBJECT //
-          $currUser   = $em
-            ->getRepository('UserBundle:User')
-            ->findOneBy(['id' => 1]);
-        // ---------------------------------------- //
-
+          $currUser   = $this->getUser();
           $formAdd    = $formAdd->getData();
           $discussion = new Discussion();
           $discussion->setContenu($formAdd['discussion']);
@@ -178,10 +174,19 @@ class ForumController extends Controller
 
     public function editDiscussionAction(Request $request, $theme, $id)
     {
+      $user       = $this->getUser();
+      $security   = $this->get('security.authorization_checker');
       $em         = $this->getDoctrine()->getManager();
       $discussion = $em
         ->getRepository('AppBundle:Discussion')
         ->findOneBy(['id' => $id]);
+
+      if (($security->isGranted('ROLE_MODO')) or ($discussion->getAuteur()->isAuthor($user))) {
+           
+      }
+      else {
+        throw $this->createAccessDeniedException('Vous n\'etes pas autorisé à accéder à cette page !');
+      }
 
       $theme = $em
         ->getRepository('AppBundle:Themes')
@@ -198,8 +203,7 @@ class ForumController extends Controller
           ->getForm();
 
       $formEdit->handleRequest($request);
-
-      if ($formEdit->isSubmitted() && $formEdit->isValid()) {
+      if ($formEdit->isSubmitted() && $formEdit->isValid() && $security->isGranted('IS_AUTHENTICATED_FULLY')) {
         try {
           $formEdit    = $formEdit->getData();
           $discussion->setContenu($formEdit['discussion']);
@@ -221,6 +225,8 @@ class ForumController extends Controller
             'id'  => $discussion->getId(),
           ));
         }
+      } else {
+        $this->denyAccessUnlessGranted('ROLE_USER', $id, 'Vous ne pouvez pas editer cet élément.');
       }
 
       return $this->render('AppBundle:Forum:editDiscussion.html.twig', array(
@@ -232,11 +238,18 @@ class ForumController extends Controller
 
     public function removeDiscussionAction(Request $request, $theme, $id)
     {
+      $user       = $this->getUser();
+      $security   = $this->get('security.authorization_checker');
       $em         = $this->getDoctrine()->getManager();
       $discussion = $em
         ->getRepository('AppBundle:Discussion')
         ->findOneBy(['id' => $id]);
 
+      if (($security->isGranted('ROLE_MODO')) or ($discussion->getAuteur()->isAuthor($user))) {
+        
+      } else {
+        throw $this->createAccessDeniedException('Vous n\'etes pas autorisé à accéder à cette page !');
+      }
       $theme = $em
         ->getRepository('AppBundle:Themes')
         ->findOneBy(['titre' => $theme]);
@@ -252,10 +265,8 @@ class ForumController extends Controller
           ->getForm();
 
       $formRemove->handleRequest($request);
-
-      if ($formRemove->isSubmitted() && $formRemove->isValid()) {
+      if ($formRemove->isSubmitted() && $formRemove->isValid() && $security->isGranted('IS_AUTHENTICATED_FULLY')) {
         try {
-
           $theme->removeDiscussion($discussion);
           $em->remove($discussion);
           $em->persist($theme);
@@ -276,6 +287,8 @@ class ForumController extends Controller
             'id'    => $id,
           ));
         }
+      } else {
+        $this->denyAccessUnlessGranted('ROLE_USER', $id, 'Vous ne pouvez pas ajouter cet élément.');
       }
 
       return $this->render('AppBundle:Forum:editDiscussion.html.twig', array(
@@ -285,6 +298,10 @@ class ForumController extends Controller
       ));
     }
 
+    /**
+     *
+     * @Security("has_role('ROLE_USER')")
+     */  
     public function profilAction($id, Request $request)
     {
       $em     = $this->getDoctrine()->getManager();
